Suppress a int64_t -> double conversion warning

+++ src/performance_counters/server/arithmetics_counter_extended.cpp
@@ -111,7 +111,7 @@ namespace hpx { namespace performance_counters { namespace server
             template <typename Accumulator>
             static double call(Accumulator const& accum)
             {
-                return boost::accumulators::count(accum);
+                return static_cast<double>(boost::accumulators::count(accum));
             }
         };
     }
